services/trade_producer/pyproject.toml
[project]
name = "trade-producer"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "loguru>=0.7.3",
    "quixstreams>=3.23.1",
    "websocket-client>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
